Move WebContextImpl to wkwebview
